==> analyses/tmb-vaf-longitudinal/04-create-violin-plots.Rmd <==
---
title: "Create TMB violin plots of tumors across multiple timepoints for thr autopsy samples in the PBTA Cohort"
author: "Antonia Chroni <chronia@chop.edu> and Jo Lynne Rokita <rokita@chop.edu> for D3B"
date: "2023"
output:
  html_notebook:
    toc: yes
    toc_float: yes
editor_options: 
  chunk_output_type: inline
---

#### Tumor evolution project 

### Data used 
In this notebook, we are using the `tmb_autopsy.tsv` file generated from the `01-preprocess-data.Rmd` script.

# Set up
```{r load-library}
suppressPackageStartupMessages({
  library(tidyverse)
})
```

# Directories and File Inputs/Outputs
```{r set-dir-and-file-names}
# Detect the ".git" folder. This will be in the project root directory.
# Use this as the root directory to ensure proper sourcing of functions
# no matter where this is called from.
root_dir <- rprojroot::find_root(rprojroot::has_dir(".git"))
scratch_dir <- file.path(root_dir, "scratch")
analysis_dir <- file.path(root_dir, "analyses", "tmb-vaf-longitudinal") 
input_dir <- file.path(analysis_dir, "input")

# Input files
tmb_autopsy_file <- file.path(scratch_dir, "tmb_autopsy.tsv")

# File path to plots directory
plots_dir <-
  file.path(analysis_dir, "plots")
if (!dir.exists(plots_dir)) {
  dir.create(plots_dir)
}

source(paste0(analysis_dir, "/util/function-create-violin-plot.R"))
source(paste0(root_dir, "/figures/theme.R"))
```

# Read in data and process
```{r read_input_files}
# Read tmb_autopsy file generated from step `01-process-data.Rmd`
tmb_autopsy <- readr::read_tsv(tmb_autopsy_file, guess_max = 100000, show_col_types = FALSE)
```

# Violin plots for each Patient case and per biospecimen_id and timepoint
```{r create-violin-plot}
# Run function
fname <- paste0(plots_dir, "/", "TMB-violin-plot-All-samples.pdf")
print(fname)
p <- create_violin_plot(tmb = tmb_autopsy)
pdf(file = fname, width = 10, height = 8)
print(p)
dev.off()
```
    
```{r create-violin-plot-exclude}
tmb_autopsy_filter <- tmb_autopsy %>%
  filter(!Kids_First_Participant_ID %in% c("PT_3CHB9PK5", "PT_6N825561"))
# Run function
fname <- paste0(plots_dir, "/", "TMB-violin-plot-All-samples-exclude.pdf")
print(fname)
p <- create_violin_plot(tmb = tmb_autopsy_filter)
pdf(file = fname, width = 10, height = 8)
print(p)
dev.off()
```    
      
# Vloin Plots based on timepoint models
There are different timepoint models, e.g., Diagnosis-Deceased or Progressive-Deceased.

```{r create-violin-plot-sample}   
timepoints_models <- unique(as.character(tmb_autopsy$descriptors))
print(timepoints_models)

for (i in seq_along(timepoints_models)) {
  print(i)
  tmb_sub <- tmb_autopsy %>%
    filter(descriptors == timepoints_models[i])
  
  # Run function
  fname <- paste0(plots_dir, "/", timepoints_models[i], "-Tmb-violin-plot.pdf")
  print(fname)
  p <- create_violin_plot(tmb = tmb_sub)
  pdf(file = fname, width = 6, height = 5)
  print(p)
  dev.off()
}
```

```{r echo=TRUE}
sessionInfo()
```
